Expose app instance for uvicorn
